fix(): brick command conf is migrated to new package
Refs NEXT_BUILDER-4421
