+++ bricks/ui-test/package.json
@@ -43,6 +43,7 @@
     "@next-core/react-element": "^1.0.16",
     "@next-core/runtime": "^1.25.7",
     "@next-core/theme": "^1.4.2",
+    "@next-shared/ui-test": "^1.0.0",
     "classnames": "^2.3.2",
     "file-saver": "^2.0.5",
     "idb-keyval": "^6.2.1",
