+++ shared/ui-test/src/data/bricks/basic-bricks.ts
@@ -1,4 +1,4 @@
-import { BrickCommandConf } from "./interfaces.js";
+import { BrickCommandConf } from "../../interfaces.js";
 
 export const basicBricks: BrickCommandConf[] = [
   {
