+++ shared/ui-test/src/data/bricks/forms.ts
@@ -1,4 +1,4 @@
-import { BrickCommandConf } from "./interfaces.js";
+import { BrickCommandConf } from "../../interfaces.js";
 
 export const formsBricks: BrickCommandConf[] = [
   {
