+++ shared/ui-test/src/data/bricks/index.spec.ts
@@ -1,0 +1,7 @@
+import { brickCommandsConf } from "./index.js";
+
+describe("brickCommandsConf", () => {
+  it("should work", () => {
+    expect(brickCommandsConf.length).toBeGreaterThan(0);
+  });
+});
