+++ shared/ui-test/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./interfaces.js";
+export * from "./data/bricks/index.js";
