+++ shared/ui-test/tsconfig.build.json
@@ -1,0 +1,4 @@
+{
+  "extends": "./tsconfig.json",
+  "exclude": ["**/*.spec.*", "**/__mocks__/", "**/__fixtures__/"]
+}
