+++ shared/ui-test/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.module.json",
+  "compilerOptions": {
+    "outDir": "./dist"
+  },
+  "include": ["./src/"]
+}
